split tests back into submodules
