+++ pgdog/src/frontend/client/query_engine/two_pc/test.rs
@@ -13,7 +13,7 @@ use super::*;
 
 #[tokio::test]
 async fn test_cleanup_transaction_phase_one() {
-    config::test::load_test();
+    config::load_test();
     let cluster = databases().all().iter().next().unwrap().1.clone();
 
     let mut two_pc = TwoPc::default();
@@ -76,7 +76,7 @@ async fn test_cleanup_transaction_phase_one() {
 
 #[tokio::test]
 async fn test_cleanup_transaction_phase_two() {
-    config::test::load_test();
+    config::load_test();
     logger();
     let cluster = databases().all().iter().next().unwrap().1.clone();
